creating a split action handler class
