+++ modelHelpers/actions/action_handler.py
@@ -33,28 +33,22 @@ class ActionHandler:
     def __init__(self, split_mode=False):
         self.split_mode = split_mode
         self.actions = self.create_actions()
-        self.action_map = ActionMap(self.actions)
+        self.action_map = self.create_action_map()
 
-        self.actions_split, self.split_action_sizes = self.create_actions_split()
-        self.action_map_split = ActionMap(self.actions_split[3])
+    def create_action_map(self):
+        return ActionMap(self.actions)
 
     def is_split_mode(self):
-        return self.split_mode
+        return False
 
-    def get_split_sizes(self):
-        return self.split_action_sizes
+    def get_action_sizes(self):
+        return [len(self.actions)]
 
     def get_action_size(self):
         """
         :return: the size of the logits layer in a model
         """
-        if not self.split_mode:
-            return len(self.actions)
-
-        counter = 0
-        for action in self.actions_split:
-            counter += len(action)
-        return counter
+        return len(self.actions)
 
     def create_actions(self):
         """
@@ -73,40 +67,8 @@ class ActionHandler:
         entirelist = list(itertools.product(*action_list))
         return entirelist
 
-    def create_actions_split(self):
-        """
-        Creates all variations of all of the actions.
-
-        controller options = [throttle, steer, pitch, steer, roll, jump, boost, handbrake]
-        :return: A combination of all actions. This is an array of an array
-        """
-
-        steer = np.arange(-1, 1.5, .5)
-        pitch = np.arange(-1, 1.5, .5)
-        roll = np.arange(-1, 1.5, .5)
-        throttle = np.arange(-1, 2, 1)
-        jump = [False, True]
-        boost = [False, True]
-        handbrake = [False, True]
-        action_list = [throttle, jump, boost, handbrake]
-        self.action_list_size = len(action_list)
-        # 24 + 5 + 5 + 5 = 39
-        button_combo = list(itertools.product(*action_list))
-        actions = []
-        split_actions_sizes = []
-        actions.append(steer)
-        actions.append(pitch)
-        actions.append(roll)
-        self.movement_actions = tf.constant(np.array(actions), shape=[len(actions), len(steer)])
-        self.combo_actions = tf.constant(button_combo)
-        self.action_list_names = ['steer', 'pitch', 'yaw', 'combo']
-        actions.append(button_combo)
-        for i in actions:
-            split_actions_sizes.append(len(i))
-        return actions, split_actions_sizes
-
     def create_controller_output_from_actions(self, action_selection):
-        if len(action_selection) != len(self.actions_split):
+        if len(action_selection) != len(self.actions):
             print('ACTION SELECTION IS NOT THE SAME LENGTH returning invalid action data')
             return [0, 0, 0, 0, 0, False, False, False]
         steer = self.actions_split[0][action_selection[0]]
@@ -158,9 +120,6 @@ class ActionHandler:
         return tf.stack(controller_option, axis=1)
 
     def create_action_label(self, real_action):
-        if self.split_mode:
-            indexes = self._create_split_indexes(real_action)
-            return self._create_split_label(indexes)
         index = self._find_matching_action(real_action)
         return self._create_one_hot_encoding(index)
 
@@ -169,27 +128,7 @@ class ActionHandler:
             return self._create_split_indexes(real_action)
         return self._find_matching_action(real_action)
 
-    def _create_split_indexes(self, real_action):
-        steer = real_action[1]
-        yaw = real_action[3]
-        if steer != yaw and abs(steer) < abs(yaw):
-            # only take the larger magnitude number
-            steer = yaw
 
-        steer_index = self._find_closet_real_number(steer)
-        pitch_index = self._find_closet_real_number(real_action[2])
-        roll_index = self._find_closet_real_number(real_action[4])
-        button_combo = self.action_map_split.get_key([round(real_action[0]), real_action[5], real_action[6], real_action[7]])
-
-        return [steer_index, pitch_index, roll_index, button_combo]
-
-    def _create_split_label(self, action_indexes):
-        encoding = np.zeros(39)
-        encoding[action_indexes[0] + 0] = 1
-        encoding[action_indexes[1] + 5] = 1
-        encoding[action_indexes[2] + 10] = 1
-        encoding[action_indexes[3] + 15] = 1
-        return encoding
 
     def _find_closet_real_number(self, number):
         if number <= -0.25:
